requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==8.4.0
flask==2.0.1
gunicorn==20.1.0
joblib